:art: move links to separate file
