+++ src/lib/links.ts
@@ -1,0 +1,10 @@
+// Socials
+export const TWITTER_LINK = "https://twitter.com/backdfund";
+export const DISCORD_LINK = "https://discord.gg/jpGvaFV3Rv";
+export const GITHUB_LINK = "https://github.com/backdfund";
+export const TELEGRAM_LINK = "https://t.me/backdchat";
+export const MEDIUM_LINK = "https://backdfund.medium.com/";
+export const NEWSLETTER_LINK = "https://backd.substack.com/welcome";
+
+// Docs
+export const DOCS_LINK = "https://docs.backd.fund/";
