+++ src/pages/not-found/NotFoundPage.tsx
@@ -4,7 +4,8 @@ import { Trans, useTranslation } from "react-i18next";
 
 import { GradientLink } from "../../styles/GradientText";
 import { useDevice } from "../../app/hooks/use-device";
-import { DISCORD_LINK, DOCS_LINK, STAKING_LIVE } from "../../lib/constants";
+import { DISCORD_LINK, DOCS_LINK } from "../../lib/links";
+import { STAKING_LIVE } from "../../lib/constants";
 
 const StyledNotFoundPage = styled.div`
   width: 100%;
